72: Adds a template file for Article nodes
I didn't want to change all the nodes on the site, but I thought that
the most important part of a blog article isn't who wrote it, but the
content of the article itself. As a result, I moved the author image and
by line to the bottom of the stack.